requirements.txt
python-telegram-bot
flask
requests
flask_sqlalchemy
python-dotenv
